fix naming; divided task1 file; divided tests;add ci run tests from script
